Reformated the structure
