--- assignment7/inc/performance.h
@@ -1,9 +1,0 @@
-#ifndef PERFORMANCE_H
-#define PERFORMANCE_H
-
-#include "player.h"
-
-float computePerformanceIndex(const Player *player);
-void computePerformanceForAll(Player players[], int playersCount);
-
-#endif
