--- assignment7/inc/search.h
@@ -1,8 +1,0 @@
-#ifndef SEARCH_H
-#define SEARCH_H
-
-#include "team.h"
-
-int searchTeamById(const Team teams[], int teamsCount, int teamId);
-
-#endif
