--- assignment7/inc/sort.h
@@ -1,10 +1,0 @@
-#ifndef SORT_H
-#define SORT_H
-
-#include "player.h"
-#include "team.h"
-
-void sortTeamsByStrikeRate(Team teams[], int teamsCount);
-void sortPlayersByPerformance(Player players[], int playersCount);
-
-#endif
