--- assignment7/src/player.c
@@ -1,23 +1,0 @@
-#include <stdio.h>
-#include <string.h>
-#include "player.h"
-
-
-void printPlayer(const Player *player)
-{
-    if (player == NULL)
-    {
-        return;
-    }
-    printf("%-5d %-20s %-12s %-12s %6d %6.2f %6.2f %6d %6.2f %8.2f\n",
-           player->id,
-           player->name,
-           player->team,
-           player->role,
-           player->totalRuns,
-           player->battingAverage,
-           player->strikeRate,
-           player->wickets,
-           player->economyRate,
-           player->performanceIndex);
-}
